Refactor: comment를 post의 하위로 패키지 이동
