--- src/main/java/goorm_runner/backend/comment/presentation/dto/CommentCreateRequest.java
@@ -1,4 +1,0 @@
-package goorm_runner.backend.comment.presentation.dto;
-
-public record CommentCreateRequest(String content) {
-}
