+++ src/main/java/goorm_runner/backend/post/application/post/PostReadService.java
@@ -1,10 +1,10 @@
-package goorm_runner.backend.post.application;
+package goorm_runner.backend.post.application.post;
 
 import goorm_runner.backend.global.ErrorCode;
-import goorm_runner.backend.post.application.exception.PostException;
-import goorm_runner.backend.post.domain.Category;
-import goorm_runner.backend.post.domain.Post;
-import goorm_runner.backend.post.domain.PostQueryRepository;
+import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.post.Category;
+import goorm_runner.backend.post.domain.post.Post;
+import goorm_runner.backend.post.domain.post.PostQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
