+++ src/main/java/goorm_runner/backend/post/domain/comment/CommentRepository.java
@@ -1,4 +1,4 @@
-package goorm_runner.backend.comment.domain;
+package goorm_runner.backend.post.domain.comment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
