+++ src/main/java/goorm_runner/backend/post/presentation/comment/CommentControllerAdvice.java
@@ -1,9 +1,9 @@
-package goorm_runner.backend.comment.presentation;
+package goorm_runner.backend.post.presentation.comment;
 
-import goorm_runner.backend.comment.domain.exception.CommentException;
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.global.ErrorResult;
-import goorm_runner.backend.post.application.exception.PostException;
+import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.comment.exception.CommentException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
