+++ src/main/java/goorm_runner/backend/post/presentation/comment/dto/CommentCreateRequest.java
@@ -1,0 +1,4 @@
+package goorm_runner.backend.post.presentation.comment.dto;
+
+public record CommentCreateRequest(String content) {
+}
