+++ src/main/java/goorm_runner/backend/post/presentation/post/PostController.java
@@ -1,11 +1,11 @@
-package goorm_runner.backend.post.presentation;
+package goorm_runner.backend.post.presentation.post;
 
 import goorm_runner.backend.global.PageMetaData;
 import goorm_runner.backend.member.application.MemberService;
 import goorm_runner.backend.member.security.SecurityMember;
-import goorm_runner.backend.post.application.PostReadService;
-import goorm_runner.backend.post.application.PostService;
-import goorm_runner.backend.post.domain.Post;
+import goorm_runner.backend.post.application.post.PostReadService;
+import goorm_runner.backend.post.application.post.PostService;
+import goorm_runner.backend.post.domain.post.Post;
 import goorm_runner.backend.post.dto.*;
 import goorm_runner.backend.postlike.application.PostLikeService;
 import lombok.RequiredArgsConstructor;
