+++ src/main/java/goorm_runner/backend/post/presentation/post/PostControllerAdvice.java
@@ -1,8 +1,8 @@
-package goorm_runner.backend.post.presentation;
+package goorm_runner.backend.post.presentation.post;
 
 import goorm_runner.backend.global.ErrorCode;
 import goorm_runner.backend.global.ErrorResult;
-import goorm_runner.backend.post.application.exception.PostException;
+import goorm_runner.backend.post.application.post.exception.PostException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
