+++ src/test/java/goorm_runner/backend/post/application/PostReadServiceTest.java
@@ -1,9 +1,11 @@
 package goorm_runner.backend.post.application;
 
-import goorm_runner.backend.post.application.exception.PostException;
-import goorm_runner.backend.post.domain.Category;
-import goorm_runner.backend.post.domain.Post;
-import goorm_runner.backend.post.domain.PostRepository;
+import goorm_runner.backend.post.application.post.PostReadService;
+import goorm_runner.backend.post.application.post.PostService;
+import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.post.Category;
+import goorm_runner.backend.post.domain.post.Post;
+import goorm_runner.backend.post.domain.post.PostRepository;
 import goorm_runner.backend.post.dto.PostCreateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
