+++ src/test/java/goorm_runner/backend/post/application/PostServiceTest.java
@@ -1,8 +1,9 @@
 package goorm_runner.backend.post.application;
 
-import goorm_runner.backend.post.application.exception.PostException;
-import goorm_runner.backend.post.domain.Post;
-import goorm_runner.backend.post.domain.PostRepository;
+import goorm_runner.backend.post.application.post.PostService;
+import goorm_runner.backend.post.application.post.exception.PostException;
+import goorm_runner.backend.post.domain.post.Post;
+import goorm_runner.backend.post.domain.post.PostRepository;
 import goorm_runner.backend.post.dto.PostCreateRequest;
 import goorm_runner.backend.post.dto.PostUpdateRequest;
 import org.junit.jupiter.api.Test;
