+++ src/test/java/goorm_runner/backend/post/presentation/PostControllerIntegrationTest.java
@@ -5,10 +5,10 @@ import goorm_runner.backend.member.domain.Member;
 import goorm_runner.backend.member.security.application.AuthService;
 import goorm_runner.backend.member.security.dto.LoginRequest;
 import goorm_runner.backend.member.security.dto.MemberSignupRequest;
-import goorm_runner.backend.post.application.PostService;
-import goorm_runner.backend.post.domain.Category;
-import goorm_runner.backend.post.domain.Post;
-import goorm_runner.backend.post.domain.PostRepository;
+import goorm_runner.backend.post.application.post.PostService;
+import goorm_runner.backend.post.domain.post.Category;
+import goorm_runner.backend.post.domain.post.Post;
+import goorm_runner.backend.post.domain.post.PostRepository;
 import goorm_runner.backend.post.dto.PostCreateRequest;
 import goorm_runner.backend.post.dto.PostUpdateRequest;
 import goorm_runner.backend.postlike.application.PostLikeService;
